+++ postgrest_py/__init__.py
@@ -1,1 +1,2 @@
-from postgrest_py.client import Client
+from postgrest_py.__version__ import __version__
+from postgrest_py.client import Client, PostgrestClient