+++ CHANGELOG.md
@@ -2,19 +2,23 @@
 
 ### Unreleased
 
-#### Features
+#### Added
 
 - Support basic authentication
 - Support stored procedures (RPC)
 
+#### Changed
+
+- Rename `Client` to `PostgrestClient` and deprecate the old name
+
 ### v0.1.1
 
-#### Bugfixes
+#### Fixed
 
 - Fix a typo in `Client.from_()`
 
 ### v0.1.0
 
-#### Features
+#### Added
 
 - Add basic features